Restyle the save button.
